Write loadtest module default & scenario config to file when writing metrics
When outputting metrics to STDIO or file, also write out the loadtest
configuration used for later reference.
When writing to a file, e.g. "MyRun.csv" also write the configuration to
a file called "MyRun.csv.config.exs" with the configuration data in
elixir literal syntax.

This solves the problem of not knowing if you ran the same loadtest
module with the same configuration. It makes sense to store the
loadtest configuration as it was used during the run alongside the
metrics that run recorded.