Add support for auto generating result output filename
If no :output option is given we'll use this.
Also allow tagging result output files with a custom string. This can be
useful when you need to run the same load tests in the same day but
against e.g. different versions (branch names) and want to tag the
metrics output files with the version name.